+++ unittest/Manager/vpg_cpp_generation_manager_test.cpp
@@ -43,7 +43,7 @@ class VPGCppGenerationManagerTest : public testing::Test
             // option for initialize source
             this->_Manager->SetWorkspace(this->GetWorkspaceSource());
             this->_Option->SetProjectType(VPGProjectType::CppComplex);
-            this->_Option->SetWorkspaceSource(L".");
+            this->_Option->SetTemplateWorkspace(L".");
             this->_Option->SetProjectName(L"VCCProjGenerator");
             this->_Option->SetProjectNameDll(L"libvpg");
             this->_Option->SetProjectNameExe(L"vpg");
@@ -57,7 +57,7 @@ class VPGCppGenerationManagerTest : public testing::Test
 
             // option for testing
             this->_Manager->SetWorkspace(this->GetWorkspaceTarget());
-            this->_Option->SetWorkspaceSource(this->GetWorkspaceSource());
+            this->_Option->SetTemplateWorkspace(this->GetWorkspaceSource());
             this->_Option->SetProjectNameExe(L"CPPProject");
             this->_Option->SetProjectNameDll(L"CPPDllProject");
             this->_Option->SetProjectName(L"ProjectName");